Rename ConfigMerger to simply `Merger`
