+++ src/Merger/Modifier/ReplaceValue.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Composer\Config\ConfigMerger\Modifier;
+namespace Yiisoft\Composer\Config\Merger\Modifier;
 
 /**
- * Object that represents the replacement of array value while performing {@see ConfigMerger::merge()}.
+ * Object that represents the replacement of array value while performing {@see Merger::merge()}.
  *
  * Usage example:
  *
@@ -24,13 +24,13 @@ namespace Yiisoft\Composer\Config\ConfigMerger\Modifier;
  *     'ids' => [
  *         2,
  *     ],
- *     'validDomains' => new \Yiisoft\Composer\Config\ConfigMerger\Modifier\ReplaceValue([
+ *     'validDomains' => new \Yiisoft\Composer\Config\Merger\Modifier\ReplaceValue([
  *         'yiiframework.com',
  *         'www.yiiframework.com',
  *     ]),
  * ];
  *
- * $result = ConfigMerger::merge($array1, $array2);
+ * $result = Merger::merge($array1, $array2);
  * ```
  *
  * The result will be
